+++ spotify_confidence/analysis/frequentist/chi_squared.py
@@ -12,25 +12,24 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import (Union, Iterable)
+from typing import (Union, Iterable, Tuple)
 
 from pandas import DataFrame
 
-from spotify_confidence.analysis.frequentist.test_classes.z_test_computer import ZTestComputer
+from spotify_confidence.analysis.frequentist.confidence_computers.chi_squared_computer import ChiSquaredComputer
 from spotify_confidence.analysis.abstract_base_classes.confidence_computer_abc import \
     ConfidenceComputerABC
 from spotify_confidence.analysis.abstract_base_classes.confidence_grapher_abc import ConfidenceGrapherABC
 from spotify_confidence.analysis.frequentist.generic_test import GenericTest
 from spotify_confidence.analysis.confidence_utils import listify
-from spotify_confidence.analysis.constants import BONFERRONI
+from spotify_confidence.analysis.constants import BONFERRONI, NIM_TYPE
 
 
-class ZTest(GenericTest):
+class ChiSquared(GenericTest):
 
     def __init__(self,
                  data_frame: DataFrame,
                  numerator_column: str,
-                 numerator_sum_squares_column: Union[str, None],
                  denominator_column: str,
                  categorical_group_columns: Union[str, Iterable],
                  ordinal_group_column: Union[str, None] = None,
@@ -39,20 +38,20 @@ class ZTest(GenericTest):
                  confidence_computer: ConfidenceComputerABC = None,
                  confidence_grapher: ConfidenceGrapherABC = None):
 
-        computer = ZTestComputer(
+        computer = ChiSquaredComputer(
             data_frame=data_frame,
             numerator_column=numerator_column,
-            numerator_sum_squares_column=numerator_sum_squares_column,
+            numerator_sum_squares_column=numerator_column,
             denominator_column=denominator_column,
             categorical_group_columns=listify(categorical_group_columns),
             ordinal_group_column=ordinal_group_column,
             interval_size=interval_size,
             correction_method=correction_method.lower())
 
-        super(ZTest, self).__init__(
+        super(ChiSquared, self).__init__(
             data_frame,
             numerator_column,
-            numerator_sum_squares_column,
+            numerator_column,
             denominator_column,
             categorical_group_columns,
             ordinal_group_column,
@@ -60,3 +59,40 @@ class ZTest(GenericTest):
             correction_method,
             computer,
             confidence_grapher)
+
+    def difference(self,
+                   level_1: Union[str, Tuple],
+                   level_2: Union[str, Tuple],
+                   absolute: bool = True,
+                   groupby: Union[str, Iterable] = None,
+                   non_inferiority_margins: NIM_TYPE = None,
+                   final_expected_sample_size_column: str = None
+                   ) -> DataFrame:
+        if non_inferiority_margins is not None:
+            raise ValueError('Non-inferiority margins not supported in '
+                             'ChiSquared. Use StudentsTTest or ZTest instead.')
+        return super(ChiSquared, self).difference(
+            level_1,
+            level_2,
+            absolute,
+            groupby,
+            None,
+            final_expected_sample_size_column)
+
+    def multiple_difference(self, level: Union[str, Tuple],
+                            absolute: bool = True,
+                            groupby: Union[str, Iterable] = None,
+                            level_as_reference: bool = None,
+                            non_inferiority_margins: NIM_TYPE = None,
+                            final_expected_sample_size_column: str = None
+                            ) -> DataFrame:
+        if non_inferiority_margins is not None:
+            raise ValueError('Non-inferiority margins not supported in '
+                             'ChiSquared. Use StudentsTTest or ZTest instead.')
+        return super(ChiSquared, self).multiple_difference(
+            level,
+            absolute,
+            groupby,
+            level_as_reference,
+            None,
+            final_expected_sample_size_column)
